backend/requirements.txt
sentence_transformers
fastapi
torch
pillow